refactor: extract VideoGen form state and payload building into useVideoGenForm (#3291)
VideoGen.jsx carried 56 useState calls, every URL-prefill effect, the mode/backend
transitions and buildGeneratePayload in one 2,100-line component, so the request
shape the client sends had no single place to read against server/routes/videoGen.js
and no second entry point could reuse it.

Form state, the ImageGen/Continue/Remix/?lora= prefill paths, the mode+backend
switches that clear stale conditioning inputs, the derived model/keyframe/IC gates
and buildGeneratePayload now live in client/src/hooks/useVideoGenForm.js. The page
keeps fetching, the SSE run pipeline, the batch queue and rendering.

Behavior-preserving: the only semantic change is that the default model is seeded
from status.defaultModel by an effect in the hook rather than inside refreshStatus's
then-callback, which lands on the same state one render later.
