+++ .changelog/NEXT.md
@@ -44,6 +44,7 @@
 - **[issue-3266] The mount guard behind issue 3264 now has exactly one implementation.** Thirteen more components and hooks still spelled that guard out by hand — correctly, so the new test passed them, but they were the copy source: reproducing issue 3264 took nothing more than copying one of them and dropping a line. Several also carried their own multi-line essay re-deriving the StrictMode hazard, which is the hazard being rediscovered per file instead of read off a hook name. All thirteen now call the shared hook, and the test fails on the hand-written form whether or not it happens to be correct.
 - **[issue-3265] Fixed a test that broke on its own once the calendar passed 2026-07-31.** An image-quota test compared against a reset timestamp baked into its fixture; once that instant passed, the code correctly ignored the stale time and the test started comparing against the wall clock, failing every run. The test now pins its own clock, and the ignore-a-past-reset path it accidentally exercised has coverage of its own.
 - **[issue-3267] Removed the unmount guard from the shared async-action hook, along with the test that could never fail.** The guard existed to avoid a "state update on an unmounted component" warning React stopped emitting in version 18, so its only remaining effect was to skip a no-op — while being the exact shape that froze every button backed by the hook under StrictMode. Its regression test asserted that the vanished warning never appeared, which stayed green even with the guard deleted; it now unmounts mid-request and asserts the action still resolves with its value.
+- **[issue-3291] The Video Gen form moved out of the page and into a hook of its own.** The page carried 56 pieces of form state and every URL-prefill, mode-transition and payload-shaping rule in one 2,100-line component, so the request the client sends had no single place to read — and no second surface could ever generate video without copying it. Form state, the ImageGen/Continue/Remix/`?lora=` prefill paths, the mode and backend switches that clear stale conditioning inputs, the keyframe/extend/audio/IC submit gates, and `buildGeneratePayload()` now live in `useVideoGenForm`, leaving the page to fetch, run, and render. Behaviour is unchanged, and the payload contract — grok versus local shapes, the empty-string sentinels, keyframe JSON, LoRA parallel arrays, and the mutually-exclusive IC reference channels — now has direct test coverage it never had before.
 - **[issue-3250] Stabilize Music Video project-selection tests under CI load.** Project-selection coverage now waits for the requested project to be available before opening its board, preventing intermittent false failures.
 - **[issue-3292] Three sprite dialogs that every suite mocked away now have tests of their own.** The fork-a-character dialog, the reference-sprite picker, and the click-to-enlarge preview each carry real logic — the submit gate, the exact payload sent when forking, search filtering by name and id, and which asset the enlarged view opens — but no test file, and every suite that rendered them swapped in a stub. The fork payload was the sharpest edge: the server only ever checked a hand-written example of it, so a mismatch between what the dialog sends and what the server accepts could have sat behind two passing suites. Both of the payloads the dialog actually builds are now asserted on the client and replayed through the real server schema. No behaviour changed.
