+++ client/src/hooks/index.js
@@ -162,5 +162,6 @@ export * from './useSyncIntegrity.js';
 export * from './useTwinEvaluationSuite.js';
 export * from './useUniverseNav.js';
 export * from './useVideoFileSrc.js';
+export * from './useVideoGenForm.js';
 export * from './useVideoGenQueue.js';
 export * from './useVoiceUiSync.js';
